+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.34.1 - 2024-05-21
+
+#### Enhancements
+- Added `use_snapshot` parameter to `Live.subscribe`, defaults to `False`
+
 ## 0.34.0 - 2024-05-14
 
 #### Enhancements
@@ -10,6 +15,9 @@
 - Fixed an issue where `batch.download` and `batch.download_async` would fail if requested files already existed in the output directory
 - Fixed an issue where `batch.download`, `batch.download_async`, and `timeseries.get_range` could use a lot of memory while streaming data
 - Fixed an issue where reusing a `Live` client with an open output stream would drop DBN records when received at the same time as the `Metadata` header
+
+##### Deprecations
+- The `start_date` and `end_date` keys in the response from `Historical.metadata.get_dataset_range` will be removed in a future release. Use the new `start` and `end` keys instead, which include time resolution
 
 ## 0.33.0 - 2024-04-16